+++ app/Main.hs
@@ -51,5 +51,5 @@ bootstrap = \case
     cgoGenMode <- parseGenMode c
     commandGen CommandGenOptions { .. }
   Init cioLicenseType cioSourcePaths -> commandInit CommandInitOptions { .. }
-  Run croSourcePaths croExcludedPaths croTemplatePaths croVariables croRunMode croDebug croDryRun
+  Run croSourcePaths croExcludedPaths croTemplateSource croVariables croRunMode croDebug croDryRun
     -> commandRun CommandRunOptions { .. }